requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
ca-certificates=2017.08.26=ha1e5d58_0
certifi=2017.11.5=py27hfa9a1c4_0
diff-match-patch=20121119=py27_0
django=1.9.5=py27_0
django-cors-headers=1.1.0=py_0
django-import-export=0.6.1=py_0
djangorestframework=3.7.3=py_0
et_xmlfile=1.0.1=py27_0
jdcal=1.3=py27_0
json-c=0.12.1=0
libcxx=4.0.1=h579ed51_0
libcxxabi=4.0.1=hebd6815_0
libedit=3.1=hb4e282d_0
libffi=3.2.1=h475c297_4
ncurses=6.0=hd04f020_2
odfpy=1.3.5=py27_0
openpyxl=2.5.0b1=py27_0
openssl=1.0.2n=hdbc3d79_0
pip=9.0.1=py27h1567d89_4
python=2.7.14=hde5916a_29
python-decouple=3.1=py_0
pyyaml=3.12=py27_1
readline=7.0=hc1231fa_4
setuptools=36.5.0=py27h2a45cec_0
sqlite=3.20.1=h7e4c145_2
tablib=0.11.5=py27_0
tk=8.6.7=h35a86e2_3
unicodecsv=0.14.1=py27_0
wheel=0.30.0=py27h677a027_1
xlrd=1.1.0=py_2
xlwt=1.3.0=py27_0
yaml=0.1.7=0
zlib=1.2.11=hf3cbc9b_2
